# openapi-spec.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Web of Trust API
  description: Providing API access to update trust scores,  retrieve trusted state and recalculate scoring.

servers:
  - url: http://localhost:20202

paths:
  /status:
    get:
      description: Return general info about server instance.
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  uptime:
                    description: uptime in seconds
                    type: integer
                  free:
                    description: free space in Mb
                    type: integer
  /trust:
    post:
      description: Set or update single trust value
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - source
                - target
              properties:
                source:
                  description: Giver of trust
                  type: string
                target:
                  description: Recipient of trust
                  type: string
                value:
                  type: number
                delta:
                  type: number
      responses:
        '200':
          description: Success set or updated a trust value
  /score:
    get:
      description: Get trusted status of target user from perspective of source.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - source
                - target
              properties:
                source:
                  description: Requesting on behalf of
                  type: string
                target:
                  description: User to return trusted state of
                  type: string
      responses:
        '200':
          description: Successfully retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  trusted:
                    type: boolean
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                type: object
                properties:   
                  message:
                    type: string
  /rescore:
    get:
      description: Recalculate scores
      responses:
        '200':
          description: Successfully retrieved
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                type: object
                properties:   
                  message:
                    type: string
  /debug:
    get:
      description: Get full state for debugging purposes.
      responses:
        '200':
           description: Successfully retrieved
           content:
             application/json:
               schema:
                 type: object
                 properties:
                   paths:
                     type: object
                   ranks:
                     type: object
                   capacities:
                     type: object
                   scores:
                     type: object